Timing logic and actor loading control are put into hooks
